high_low_bot/requirements.txt
python-telegram-bot
python-binance
python-dotenv